Add MIDI in/out menus; protect against Default bug
- Added trait helpers to make it easier to notice when a recv fails.
- Instead of hardcoding a MIDI input, make it dynamic. This wasn't too hard!

I keep getting bitten by the desire to use ..Default::default(). In this case, I
had ActorSystemApp::new() create a struct by filling in certain fields and then
calling ..Default::default(). But both new() and the default() implementation
created a ServiceManager. So we ended up with one service that got dropped
quickly, leading to all sorts of confusion (dead channels).

The problem is that I have been thinking of ..Default::default() as a macro that
magically figures out the fields that are missing and then fills them in. But I
think it *calls* default(), doing all the work in that body, and then copying
over the needed fields. So if I want to use ..Default::default(), I need to
limit it to cases where the init function does only idempotent work (or go
farther and make a rule that all such functions must be idempotent).